Betty fix: split _printf into handlers; parentheses on _putchar; Task 0 clean

+++ _putchar.c
@@ -4,9 +4,9 @@
  * _putchar - writes a single character to stdout
  * @c: character to print
  *
- * Return: 1 on success, or -1 on error.
+ * Return: 1 on success, -1 on error.
  */
 int _putchar(char c)
 {
-	return (int)write(1, &c, 1);
+	return (write(1, &c, 1));
 }
